Battle file cleanup & documentation (#8903)

+++ include/battle_dynamax.h
@@ -7,8 +7,8 @@ bool32 CanDynamax(u32 battler);
 bool32 IsGigantamaxed(u32 battler);
 void ApplyDynamaxHPMultiplier(struct Pokemon *mon);
 void ActivateDynamax(u32 battler);
-u16 GetNonDynamaxHP(u32 battler);
-u16 GetNonDynamaxMaxHP(u32 battler);
+u32 GetNonDynamaxHP(u32 battler);
+u32 GetNonDynamaxMaxHP(u32 battler);
 void UndoDynamax(u32 battler);
 bool32 IsMoveBlockedByMaxGuard(enum Move move);
